DUME Sentry
Adds a "D.U.M.E Sentry Kit" to the tower inventory. It has a single ability to assemble it into a Sentry Gun Paragon. When assembled, all sentry guns on screen are sacrificed to determine it's Paragon Power.

The Paragon Power is determined by the number of sentry guns, their tier, pops, and a small % of pops of their parent Engineer.