Added Training Tower
The training monkey can spawn a "Dummy Boss" which has infinite health and does not drain any lives when leaked. All damage dealt to the Dummy Boss is tracked on the training monkey's damage counter and can be reset. Intended for tracking damage for testing strategies or just comparing damage numbers